+++ spring-core/src/main/java/org/springframework/core/annotation/AbstractAliasAwareAnnotationAttributeExtractor.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2002-2016 the original author or authors.
+ * Copyright 2002-2019 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -64,7 +64,7 @@ abstract class AbstractAliasAwareAnnotationAttributeExtractor<S> implements Anno
 		this.annotationType = annotationType;
 		this.annotatedElement = annotatedElement;
 		this.source = source;
-		this.attributeAliasMap = AnnotationUtils.getAttributeAliasMap(annotationType);
+		this.attributeAliasMap = InternalAnnotationUtils.getAttributeAliasMap(annotationType);
 	}
