+++ spring-core/src/main/java/org/springframework/core/annotation/SynthesizingMethodParameter.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2002-2016 the original author or authors.
+ * Copyright 2002-2019 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -103,7 +103,6 @@ public class SynthesizingMethodParameter extends MethodParameter {
 	protected Annotation[] adaptAnnotationArray(Annotation[] annotations) {
 		return AnnotationUtils.synthesizeAnnotationArray(annotations, getAnnotatedElement());
 	}
-
 
 	@Override
 	public SynthesizingMethodParameter clone() {
